requirements.txt
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
